Рефакторинг_2
+ обработка задания отделена от первичного создания интерфейса
+ оптимизированы CheckCheck и RadioCheck (List и имена кнопок)
