--- lb1/src/by/bs/swt/Main.java
@@ -1,8 +1,0 @@
-package by.bs.swt;
-
-public class Main {
-	public static void main(String[] args) {
-		Linker num = new Linker();
-		num.run();
-	}
-}
